persist: move the Snapshot{Ext} traits to the indexed module
