+++ src/persist/src/indexed/future.rs
@@ -19,8 +19,7 @@ use timely::progress::{Antichain, Timestamp};
 use crate::error::Error;
 use crate::indexed::cache::BlobCache;
 use crate::indexed::encoding::BlobFutureMeta;
-use crate::indexed::BlobFutureBatch;
-use crate::persister::Snapshot;
+use crate::indexed::{BlobFutureBatch, Snapshot};
 use crate::storage::{Blob, SeqNo};
 use crate::Data;
