+++ src/persist/src/indexed/trace.rs
@@ -22,8 +22,7 @@ use timely::progress::{Antichain, Timestamp};
 use crate::error::Error;
 use crate::indexed::cache::BlobCache;
 use crate::indexed::encoding::BlobTraceMeta;
-use crate::indexed::BlobTraceBatch;
-use crate::persister::Snapshot;
+use crate::indexed::{BlobTraceBatch, Snapshot};
 use crate::storage::Blob;
 use crate::Data;
